+++ source/renderer/assets/codemirror/zettlr-plugin-render-citations.js
@@ -15,7 +15,7 @@
   // Should match everything permittible -- first alternative are the huge
   // blocks, second alternative are the simple @ID-things, both recognised by
   // Pandoc citeproc.
-  var citationRE = /\[([^[\]]*@[^[\]]+)\]|(?<=[^\\])@([a-z0-9_:.#$%&\-+?<>~/]+)/gi
+  var citationRE = /(\[[^[\]]*@[^[\]]+\])|(?<=[^\\])@([a-z0-9_:.#$%&\-+?<>~/]+)/gi
   var citeMarkers = [] // CiteMarkers
 
   CodeMirror.commands.markdownRenderCitations = function (cm) {
@@ -48,50 +48,7 @@
 
       // Run through all links on this line
       while ((match = citationRE.exec(line)) != null) {
-        let realCitations = []
-        if (match[1]) {
-          let cite = match[1].split(';') // First extract all citations in this thing
-          for (let c of cite) {
-            // It could be that the user just ended his/her citation with a ;
-            if (c === '') continue
-            // Loop through all array items and deconstruct them. They can look like this:
-            /*
-            [see @doe99, pp. 33-35; also @smith04, chap. 1]
-            [see @doe99, pp. 33-35; also @smith04, chap. 1]
-            [@doe99, pp. 33-35, 38-39 and *passim*]
-            [@smith04; @doe99]
-            [-@smith04]
-             */
-            let myCitation = {}
-            myCitation['prefix'] = c.split('@')[0] // Now in pre there is the prefix
-            myCitation['suppress-author'] = c.indexOf('@') > 0 && c[c.indexOf('@') - 1] === '-' // Should the author be omitted?
-            if (myCitation['suppress-author']) myCitation['prefix'] = myCitation['prefix'].substr(0, myCitation['prefix'].length - 2)
-            // Now we should have an ID. The list of characters that is permitted
-            // inside a citation key (e.g. the id) is taken from
-            // http://pandoc.org/demo/example19/Extension-citations.html
-            // This does not work currently: /[^a-z0-9_:.#$%&-+?<>~/]/
-            let suffixWithID = c.split('@')[1].split(/[,; ]/)
-            myCitation['id'] = suffixWithID.shift()
-            // We may have some information loss here, but it's all about previewing, not being perfect
-            myCitation['suffix'] = suffixWithID.join(' ')
-            // Last but not least, try to find if in the suffix there is a locator
-            if (myCitation['suffix'].length > 0) {
-              // Allow for maximal 99999 pages inside a publication.
-              let m
-              if ((m = /\d{1,5}(-\d{1,5})?/.exec(myCitation['suffix'])) != null) {
-                myCitation['locator'] = m[0]
-                myCitation['suffix'] = myCitation['suffix'].replace(m[0], '') // Remove locator from the rest of the suffix
-              }
-            }
-            myCitation['prefix'] = myCitation['prefix'].trim()
-            myCitation['suffix'] = myCitation['suffix'].trim()
-            realCitations.push(myCitation)
-          }
-          // END IF MATCH[1]
-        } else {
-          // MATCH[2] -- simple IDs. This is so silly.
-          realCitations.push({ 'id': match[2] })
-        }
+        let citation = match[1] || match[2]
 
         // Now get the precise beginning of the match and its end
         let curFrom = { 'line': i, 'ch': match.index }
@@ -117,9 +74,7 @@
         let span = document.createElement('span')
         span.className = 'citeproc-citation' // citations
         // The text content will be updated automatically based upon the ID
-        span.textContent = match[0]
-        // The attribute will be taken by the citation updater to update the citations
-        span.setAttribute('data-citeproc-cite-item', JSON.stringify(realCitations))
+        span.textContent = citation
         // Apply TextMarker
         try {
           let textMarker = cm.markText(